remove roles allowed because of possible bug

+++ src/main/java/org/disx/DisxResource.java
@@ -3,10 +3,8 @@ package org.disx;
 import java.util.List;
 import java.util.Set;
 
-// import io.quarkus.security.Authenticated;
 import io.quarkus.security.Authenticated;
 import io.quarkus.security.identity.SecurityIdentity;
-import jakarta.annotation.security.RolesAllowed;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
 import jakarta.ws.rs.DELETE;
@@ -23,6 +21,9 @@ public class DisxResource {
 
     @Inject
     DisxService DisxService;
+
+    @Inject
+    SecurityIdentity securityIdentity;
 
     @GET
     public List<Disx> getDisxs() {
@@ -56,9 +57,13 @@ public class DisxResource {
 
     @DELETE
     @Path("/{id}")
-    @RolesAllowed("admin")
+    @Authenticated
     public Response deleteDisxById(@PathParam("id") Long id) {
-        DisxService.deleteDisx(id);
-        return Response.status(Response.Status.NO_CONTENT).build();
+        Set<String> s = securityIdentity.getRoles();
+        if (s.contains("admin")) {
+            DisxService.deleteDisx(id);
+            return Response.status(Response.Status.NO_CONTENT).build();
+        }
+        return Response.status(Response.Status.UNAUTHORIZED).build();
     }
 }
